requirements.txt
openai>=1.40.0
python-dotenv>=1.0.1
requests>=2.31.0
beautifulsoup4>=4.12.2
lxml>=5.2.2
numpy>=1.26.4
python-telegram-bot>=20.7
